requirements.txt
faster-whisper==1.0.3
openai>=1.40.0
gspread==6.1.2
google-auth==2.33.0
python-dotenv==1.0.1
tenacity==8.3.0
tqdm==4.66.4
